SIN-3 :fire: Conectando proyecto, base de datos y congfig python.
