+++ project/__init__.py
@@ -2,7 +2,8 @@ from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 import time
 
-from .routers import user_router, tank_router, opc_router, auth_router, llenadera_router, bitacora_router, reloj_router, barreras_router
+from .routers import user_router, tank_router, auth_router, llenadera_router, bitacora_router, reloj_router
+#barreras_router, opc_router
 
 from .database import User
 from .database import TanksEntry
@@ -19,8 +20,8 @@ from .database import Folio
 from .database import TankExit
 from .database import database as connection
 
-from .opc import OpcServices
-from .daemon import OpcDaemon
+#from .opc import OpcServices
+#from .daemon import OpcDaemon
 
 app = FastAPI(
     title='SCADA-IRGE',
@@ -46,12 +47,12 @@ app.add_middleware(
 
 app.include_router(user_router)
 app.include_router(tank_router)
-app.include_router(opc_router)
+#app.include_router(opc_router)
 app.include_router(auth_router)
 app.include_router(llenadera_router)
 app.include_router(bitacora_router)
 app.include_router(reloj_router)
-app.include_router(barreras_router)
+#app.include_router(barreras_router)
 
 @app.on_event('startup')
 def startup():
@@ -73,7 +74,7 @@ def startup():
         TankEntry,
         TankExit,
     ])
-    
+    return
     if OpcServices.conectarOPC():
         print('conectado')
     else:
