+++ project/routers/__init__.py
@@ -3,6 +3,6 @@ from .tanks import router as tank_router
 from .llenaderas import router as llenadera_router
 from .bitacora import router as bitacora_router
 from .reloj import router as reloj_router
-from .barreras import router as barreras_router
-from .opc import router as opc_router
+#from .barreras import router as barreras_router
+#from .opc import router as opc_router
 from .auth import router as auth_router
