+++ requirements.txt
@@ -8,7 +8,7 @@ fastapi==0.70.1
 gunicorn==20.1.0
 h11==0.12.0
 idna==3.3
-mysqlclient @ file:///C:/Users/SCA/Documents/python-proyects/opc-apy/mysqlclient-1.4.6-cp38-cp38-win32.whl
+mysqlclient==2.1.1
 OpenOPC-Python3x==1.3.1
 peewee==3.14.8
 pycparser==2.21
@@ -23,5 +23,5 @@ serpent==1.40
 six==1.16.0
 sniffio==1.2.0
 starlette==0.16.0
-typing-extensions==4.0.1
+typing_extensions==4.0.1
 uvicorn==0.16.0
